move the slasher mock
