+++ beacon-chain/rpc/prysm/v1alpha1/slasher/BUILD.bazel
@@ -26,7 +26,7 @@ go_test(
     ],
     embed = [":go_default_library"],
     deps = [
-        "//beacon-chain/slasher:go_default_library",
+        "//beacon-chain/slasher/mock:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//testing/require:go_default_library",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
