+++ beacon-chain/slasher/BUILD.bazel
@@ -10,7 +10,6 @@ go_library(
         "helpers.go",
         "log.go",
         "metrics.go",
-        "mock_slashing_checker.go",
         "params.go",
         "process_slashings.go",
         "queue.go",
@@ -76,6 +75,7 @@ go_test(
         "//beacon-chain/core/signing:go_default_library",
         "//beacon-chain/db/testing:go_default_library",
         "//beacon-chain/operations/slashings/mock:go_default_library",
+        "//beacon-chain/slasher/mock:go_default_library",
         "//beacon-chain/slasher/types:go_default_library",
         "//beacon-chain/state/stategen:go_default_library",
         "//beacon-chain/sync/initial-sync/testing:go_default_library",
