+++ beacon-chain/slasher/mock/mock_slashing_checker.go
@@ -1,4 +1,4 @@
-package slasher
+package mock
 
 import (
 	"context"
